Revert notification stream changes (#13)
It seems like the new notification stream strategy is causing crashes on the application that are not easily debuggable. The crash are random and sometime causing the phone to reboot.
This commit reverts the notification stream processing to the previous state.
We will continue the work on the new notification stream strategy on a separate branch and merge it back to develop only when stable enough.