Add message to hiddenInConversation relation instead of deleting it (#11)
* Add message to hiddenInConversation relation instead of deleting it

There is a possible race condition in which a message has been deleted when a second part (e.g. link preview or medium asset) arrives.
In that case the message would have been re-added to the conversation if we deleted it in the managed object context.
To avoid this issue we now add the message to the hiddenInConversation relation on it's conversation and remove it from it's visibleInConversation relation.

* Update `hasBeenDeleted` computed property getter

* Bump dependencies