Unify UIMenuController item order to make `delete` always appear last (#104)
* Unify UIMenuController item order to make `delete` always appear last

From the Apple documentation:  "Custom items appear in the menu after any system menu items." This means that the `-delete:` selector can not be used, but a custom menu item has to be added if we want to guarantee that the delete item is displayed last (after other custom items like for example `save`). Because the `-delete:` can not be used we now use `-deleteMessage:` as custom selector and spend a custom menu item to the menu controller.

* Revert conforming ConversationCell to UIKeyInput