New address book upload specs (#29)
* Implement new address book upload specs

Uploading the address book is in fact a search. The backend does not store address book, so if I upload only once, I will never get any match for users that join in the future. For this reason, I need to constantly search using the contacts I have in the address book.
Code changes:
- Ported most of the code to Swift
- Removed upload of "self" card. This is not required anymore from the backend
- Do not check if the address book changed since last upload. Now we need to upload even if the list of contacts did not change
- Some parts still covered by TODOs

* Add uploading in batches

* Add copyright

* Update tests

* Rename files for consistency

* Fix comments and code structure