Fix searching with address book when some contacts do not have valid … (#32)
* Fix searching with address book when some contacts do not have valid email / phone

We get the information on how many AB contacts are available directly from the AB (because we don't want to parse/normalize ALL of them just to know how many there are, it might take a long time). However some of these contacts might actually have invalid email or phone number, so the actually number of contacts that we can upload might be lower.
Before this PR, we were relying on the number of contacts in the AB, and not the number of "uploadable" contacts, to know if we reached the last batch of the search and we should restart from the first. But in case of non valid contacts, we would think that we never reach the end.
Now we correctly check if we reached the last "upload able" contact by checking if we have less contacts in the batch that we could have, and if that's the case, reset the batch counter so that we reupload the first one the next time.

Also, when trying to upload a new batch, it could be that the computed batch is empty - because the previous batch stopped exactly at the last "uploadable" contact. If this is the case, and we are trying to upload this (empty) batch, immediately reupload the first one instead.

* Removed duplicated comment

* Fixed comment