Port event decoder branch to new repository (#28)
* Port branch from old repository

This PR ports the branch `feature/add-event-decoder-ZIOS-5318` to the new repository.
This adds a separate database on the sync-engine which is used to store received update events before they have been decrypted.
This database ensures the decrypted events are not lost in case of a crash before processing them (as they can only be decrypted once).

* Inject shared app group container url

* Merge branch 'develop' into feature/add-event-decoder-ZIOS-5318

# Conflicts:
#	Cartfile.resolved
#	Tests/Source/E2EE/UserClientRequestFactoryTests.swift
#	Tests/Source/UserSession/BackgroundAPNSPingBackStatusTests.swift

* Bump ZMCDataModel

* Update with comments from PR review