Do not clear the sender when removing a message caused by an edit
We previously removed the sender remote identifier from a message when calling `[ZMMessage removeMessage]` on it, but in case of an edit we still need this information to update the new message with the correct sender identifier. This lead to an edited message not displaying the sender name / image.
The `[ZMMessage removeMessage]` has been renamed to `[ZMMessage removeMessageClearingSender:]` and the parameter indicates weather the sender information should be removed.
This also fixes some flaky test for asset client messages.