Move away from system API and fallback to moving files manually
The system API does not move the external data located in `.\(database-name)_SUPPORT` and the store to be moved needs to be opened (and thus migrated before moving it), the documentation also states both stores will be open-end and all objects copied over (which makes sense as this method is also used to change the store's type), which might lead to more memory being used than necessary in our case. As we only want to move the store once on startup we now fall back to the way we moved the store before and also move the auxiliary data blob files.