Likes/reactions support (#143)
* Initial timestamp logic update

* WIP

* Comments cleanup

* Corrected license

* Updated resend button

* Removed unused code

* Updated tests

* Corrected file messages click logic

* Like animation

* Updated like animation

* Corrected like button

* Update for review

* Fixed tests

* Removed unused code

* Update for new data model

* WIP likes

* Use sync engine from branch

* Fixed tests

* Additional like actions

* WIP

* Bump wire-ios-cryptobox 3.1.2

* WIP (dirty)

* Bump components

* Wip likers list

* WIP

* Fixes for likes UI in conversation

* Logic to swap timestamp with likers list

* Resend message copy

* Reactions names logic

* Reactions API support

* Visual bugs fixes

* Do not show reactions when there is none

* Updated tests

* Update to new SE API and bump dependencies

* Visual fixes

* Update Cartfile.resolved

* Image message cell actions

* Fixed UI tests

* Visual fixes

* Fix reactions not updating when a user first likes a message

The reactionsChangeInfo is `nil` in case the user has not reacted to a message before, this is caused by the way the observation internally works. The Message observer observers its reactions relation and adds a `ReactionObserver` to added reactions, this means that this `ReactionObserver` can not fire in case this reactions has not been added before. Because of this issue the UI should always check the `reactionsChanged` flag instead of the `reactionsChangeInfo` directly. We might address this issue on the data-model side by either creating a `ReactionChangeInfo` manually or making the `reactionsChangeInfo` private.

* Updated test fixtures

* Fixed copyright header

* Fixed typo in ZMMessageReaction

* Bump wire-ios-protos 6.4.1

* Android compatibility

* Corrected message update