Fix multithreading violation in EventDecoder (#53)
* Perform event decoding on correct queue

* Ensure the next batches is only processed after the previous one has been deleted.

Add the `eventMOC`'s dispatchGroup to the tests.

* Add the `syncMOC` dispatchGroup to the `eventMOC` to ensure we wait for it in all tests

* Revert disabling sending delivery status confirmations

* Refactor `EventDecoder` to process all stored events

The event processing in the `EventDecoder` has been refactored and another multithreading violation has been removed.
Before this change there was an optimization to avoid refetching the just inserted `StoreUpdateEvent`'s, this optimization has been removed as it made the code harder to reason about.