Fix failing test
Checking that the payload is equal to the original message should not have been done in the first place. Every time you encrypt (the same message), you will get a different result as the keys are rotated. It was returning the same because, missing a session with the receiver, it could not encrypt anything and was basically returning an empty message with only the ID of the sender.

A proper fix would be to decode the message that is sent as if we were the receiver, and inspect it. So far we only check that the payload is not nil which is not reliable at all.