Add first ephemeral messages UI (#298)
* Add first ephemeral sending implementation

Text message sending (without link previews) is working, the placeholder images will need to be generated for images and link previews.

* Add ephemeral colored separator to InputBar

* Add obfuscated states for text, location, image, file / audio and video message cells

* Add DestructionCountdownView with tests

* Add missing project entries

* Update dependencies

* Add countdown timer view for ephemeral messages

* Add one minute ephemeral timeout option

* Fix test compilation

* Update ephemeral messaging copy (#294)

The cursor hint (& feature name) is TIMED MESSAGE to correspond with the hourglass icon that accompanies it, and the timer setting copy helps explain that timed messages are temporary.

* Fix wrong target membership

* Move display link in super class, add obfuscation views to all cells

* Add test for obfuscated text message cell

* Use new ephemeral color and add test for obfuscated cells

* Add comments from PR review

* Add comment to UIPickerView hack

Design agreed to leave this hack in.