Use ephemeral color for send button when in ephemeral mode (#302)
Extract button hide logic into a separate class to declutter the `ConversationInputBarViewController` a bit.
Add a flag to `IconButton` to adjust the backgroundImage when highlighted.
The hourglass is now only shown when in not in ephemeral mode and no text is in the input bar, the destruction time indicator is shown when no text is in the input bar but when in ephemeral mode.