Fix a regression which lead to the draft message text being reset (#308)
The conversation draft message text was reset for all input bar state changes, but it should only be reset in case the user finished to cancelled an edit.
This also fixes an issue in which the right accessory view visibility was not updated in some cases (e.g. after finishing an edit).