Add asset v3 preview image download request strategy
This adds a new strategy which is only used to download preview images for files uploaded using the v3 endpoint.
The reasoning not to put this logic in the v3 file download strategy is as follows:

* There are multiple things we want for file downloads but not for the image preview, like progress etc.
* The AssetDownloadStrategy already has to deal with v2 and v3 assets, increasing this to preview images makes the code hard to reason about and unnecessarily bloats this class.

Update tests for a new data model with a asset client message version identifier.