Add whitelist to AssetV3PreviewDownloadRequestStrategy and fix predicate
The `AssetV3PreviewDownloadRequestStrategy` used a wrong predicate only downloading the preview when the assets `transferState` was `.downloading`, instead we want to whitelist the image download when the UI calls `requestImageDownload`.
Thus this PR changes the implementation to use a `ZMDownstreamObjectSyncWithWhitelist` and listen to image download notifications as well as  adjusts the predicate for objects that need to be downloaded.