Add support to receive assets sent using the /assets/v3 endpoint (#17)
* Add initial logic to download assets from the v3 endpoint

* Add comments and rename test helper

* Add asset v3 preview image download request strategy

This adds a new strategy which is only used to download preview images for files uploaded using the v3 endpoint.
The reasoning not to put this logic in the v3 file download strategy is as follows:

* There are multiple things we want for file downloads but not for the image preview, like progress etc.
* The AssetDownloadStrategy already has to deal with v2 and v3 assets, increasing this to preview images makes the code hard to reason about and unnecessarily bloats this class.

Update tests for a new data model with a asset client message version identifier.

* Use assetId as cache key for v3 assets

* Add version 3 checks to download strategy predicates

* Extract v3 asset downloading in separate request strategy

* Add whitelist to AssetV3PreviewDownloadRequestStrategy and fix predicate

The `AssetV3PreviewDownloadRequestStrategy` used a wrong predicate only downloading the preview when the assets `transferState` was `.downloading`, instead we want to whitelist the image download when the UI calls `requestImageDownload`.
Thus this PR changes the implementation to use a `ZMDownstreamObjectSyncWithWhitelist` and listen to image download notifications as well as  adjusts the predicate for objects that need to be downloaded.

* Use correct cache key

* Fix asset v3 download cancellation and add test

* Add cache changes ported from v3 sending branch

* Remove unneeded code

* Bump dependencies