Return first suggested handle when receiving 404 checking available handles (#151)
When bulk checking handle availabilities we were only acknowledging a 200 response as potentially available (parsing the response to check for which handles we did not receive a payload).
As the used endpoint is not a specific search endpoint but the usual endpoint to get user profiles we also need handle a 404 where all handles are available.
If none of the handles belongs to an existing user the endpoint returns a 404 NOT FOUND and we now use the first suggested handle.