Feature/usernames conversation menu (#450)
* Split connection view in incoming and outgoing classes

This splits `UserConnectionView` in two separate classes, `UserConnectionView` and `IncomingConnectionView`. While `UserConnectionView` does not contain any controls anymore (the cancel request and archive button have been moved into a separate class called `OutgoingConnectionBottomBar`) and is shown at the top of 1-on-1 conversations, the `IncomingConnectionView` is used to show incoming requests and has a ignore and accept button. The controls for outgoing connection requests have been moved into a separate control to make the header more reusable and fix layout issues for outgoing requests.

* Add functionality to use an arbitrary UIView as title view in ActionSheets

* Add missing @-sign to invitation copy

* Adjust search bar copy

* Add comments from PR review