Use new data-model/cryptobox API and do sane things in tests
In the test, because we did not have two encryption contexts, we were using one single encryption context for two users (self and the other test user). Basically one instance was generating prekeys, then the same instance was using that prekey to establish a session with itself. This is risky because if by any chance those two (self and other client) end up having the same session identifier, then they will write to the same session file, messing it up (after one write, the other side won't be able to use it anymore to decode). There is also always a risk of threading issues even in case of distinct session identifiers, as cryptobox per se is not thread safe.

Now in tests we create separate cryptobox instances for each test user, so no more clashes.