Fix interrupting audio session when receiving a 2nd call (#270)
We were reset:ing the audio session on performing the end call action. This is
a problem when you are already in a call and receives a 2nd call. Declining that
second call would interrupt audio session and leave you muted for the rest of call.

A better for for reset:ing the audio session is in the `didDeactivateAudioSession`
which is called after a call is over