Fix stack initialization (#333)
* Update project and components

* Singleton for initialisation and tear down of CoreData stack

The `StorageStack` singleton should be used to create the stack, using the `createManagedObjectContextDirectory:…` method. It will asynchronously return a `ManagedObjectContextDirectory` that contains all the necessary `NSManagedObjectContext`. It will internally take care of migration, waiting for the filesystem to be unlocked if awaiting first unlock, and any other internal operation.

To tear down the stack, call `StoreStack.reset()` after releasing any reference to the managed object contexts. After calling this method, the stack is ready to be initialised again with  `createManagedObjectContextDirectory:…`.

* Fixed headers

* Mark migration as Optional, add storeURL properties (#325)

* Add nullability annotations

* Move store exists method to StorageStack (#331)

* Fixes for tests (#332)

* WIP

* WIP

* fix ZMTestSession

* Fix tests

* Bump components

* Fix DataBaseTest

* Fix migration tests

* Fix ZMManagedObjectGroupingTests

* Fix DatabaseBaseTest

* Fix DatabaseMovingTests

* Fix tests

* delete tests