Move LocalStoreProvider creation into static function
The `LocalStoreProvider` object is not retained in the `SessionManager` (neither should or was `self`in any of the involved completion handlers), but only used to invoke the `create...` method method on it which in turn will pass a complete instance of `LocalStoreProvider` in its completion handler. This instance is then passed to the user session init. This API design had a couple of shortcomings which are addressed in this PR.

Before this change an instance of `LocalStoreProvider` was created with its `contextDirectory` set to `nil`. After invoking the async `create...` method on that object the completion handler was called passing on the object itself with the directory set. I consider this as not the best API design and converted the create method to be `static`. The completion handler still hands over a `LocalStoreProvider` instance, but now there is a guarantee that every instance will always hold a `ManagedObjectContextDirectory`.