Fix: notification for change is not fired in certain cases (#350)
* Fix: notification for change is not fired in certain cases

CoreData has a concept of “Uniquing” which prevents creation of multiple NSManagedObject instances pointing to the same underlying CoreData record, but it does not guarantee the same behavior for faulted objects. When the last reference to the faulted object dissapears CoreData is free to re-create the object again when needed. This poses a problem when used together with NotificationCenter because it captures the “sender” object weakly. If the sender gets deallocated the notification is not delivered to the subscriber.

To work around this issue we keep a strong reference to the object until the token is alive.

* Spelling