Re-organize folder structure (#358)
* Re-organize folder structure

The various data we persist for an account is scatterd in different folder so
we decided we need clean it up and put everything inside one folder per account.

We store our data in a shared container. A shared container has the same structure
as the main bundle container. We want keep all the account related data in a folder
named by the account-id (UUID) so that it's easy to delete.

```
Accounts/
    <UUID>                    -- JSON file storing basic account information
    ...

AccountData/
    <UUID>/
        otr/                  -- folder containing cryptobox sessions
        store/                -- persistence store for messages etc.
        events/               -- persistence store for unprocessed events
        drafts/               -- persistence store for message drafts
        sharedObjectStore     -- booking keeping to sync changes between share extension and main app
    ...

Library/Caches/
    wire-account-<UUID>/
        <PIN cache folders>   -- store images and other data we can download on demand.
```

* Remove commented out code

* Update comment