Xcode 9 support (#583)
* Be more precise about what keypath we mean

* Optionality changes

* Pinning data model branch

* Pinned components

* Updated circle config

* Removed traces of old activity things

* wip

* Bump components

* Bump components

* Removed romefile

* Fix test compilation

* Fix: test failed because it had different case for user identifier

* Bump components

* Workaround: OCMock corrupts memory when running sync strategy tests

ZMSyncStrategyTests uses a huge amount of mocking. To check for some sync states the ApplicationStatusDirectory is mocked together with some statuses. However after building with Xcode9 it somehow corrupts memory. The first variable of ApplicationStatusDirectory is set to pointer value 0x1 and any calls it to it crash the test suite. Changing the order of properties corrupts the apnsConfirmationStatus pointer value, but this is not used so test passes.

* Mocking of NSManagedObject crashes on Xcode9

* Revert to regular resource branch

* Bump components

* Don’t cache carthage folder

* Fixed warning

* Making sure we pin AVS version correctly

* Fix warnings