Fix crash because of wrong change participant change update (#592)
* Fix crash because of wrong change participant change update

Problem
When starting a call, the participants snapshot will be created when the first
person joins the call. This however did not create a change notification. When
the second person joins the first change set will not include first person.
This wasn't a problem before since we created the participants collection view
after the first person had joined the call.

Solution
Also create a change notification for the first person joining the call.

* Update test