don’t require non active app state to handle call action (#596)
* don’t require non active app state to handle call action

on iOS 11, tapping a call notification would call this method with an active app state, meaning we would not create and process the pending notifcation. Allowing an active state is acceptable because this method is only called when tapping a system notifcation, which can only be done when the app is not in the foreground.

* remove uneccessary test

now we don’t require a non active app state when handling local notifications, this test was breaking because “showConversationListForUserSession” was being called. However, after some discussion with Jacob, we concluded that this test is outdated and no longer reflects a realistic situation because we wouldn’t have a user session if we were not already logged in.