Fix for double system message (#386)
Clean up ZMSystemMessage. No logical changes.

---

Problem:
When a user changes a conversation name or add/remove a participant we currently will insert
two system message for this event. This happens because the change event arrives twice, first
in the response of POST response and second in the notification stream.

Solution:
Add a check before inserting these system messages for if the change has already been applied.