Fix for double system message (#605)
* Fix for double system message

ZMConversationTranscoder has taken over the task of inserting system messages.

NOTE: also noticed that processDownloadedEvents was never called anywhere and
that therefore all events are considered "live".

---

Problem:
When a user changes a conversation name or add/remove a participant we currently will insert
two system message for this event. This happens because the change event arrives twice, first
in the response of POST response and second in the notification stream.

Solution:
Add a check before inserting these system messages for if the change has already been applied.

* Remove OCMock usage and nil out objects in tearDown