Add workaround to fix empty list when opening push (#1683)
When leaving the application with the settings screen being visible last and opening a push notification to get into the app again, we try to dismiss all presented view controllers before showing the selected conversation. The settings screen is presented using a custom view controller transition `SwizzleTransition` which is presented over the current context and applies some transforms and alpha modifications to the conversation list.

For some reason though, the `transitioningDelegate` is not being queried to provide an `UIViewControllerAnimatedTransitioning` object when the view controller is being dismissed without animation. I would have assumed the delegate being queried and the transition contexts property `isAnimated` evaluating to false, so that the pre-presentation state can be restored. I haven't figured out why this is not the case here and will continue investigating.

This PR adds a fix restoring the correct alpha and transform of the `conversationListViewController's` view when dismissing modals non-animated.