Migrate to Xcode 10 and Swift 4.2 (#2666)
* Update to Xcode 10

* Bump components

* Update for iOS 12 SDK

* Fix tests

* Update to Swift 4.2

* Fix compilation error

* Fix link style attribute crash

Incorrect type inference treated `[]` as an Array, instead of an empty `NSUnderlineStyle` option set.

* Fix failing settings test

XCTAssertNil could not infer the type of the value returned from the Settings, and couldn’t check for nullability.

* Fix flaky string matching

Due to new changes in Hashable, the order of keys in a dictionary is no longer predictable, so we have to update the tests accordingly. This does not impact actual code that uses the function, as we were not making assumptions about the order.

* Bump released components

* Bump components

* Updated snapshots

* Removed old snapshots and re-recorded tests

* Fix link text color

* Revert bundle ID change

* Fix flaky test