ZIOS-10730: Fix email login prefill after logging in with phone (#2806)
* Prefill the email for needsPasswordToRegisterClient error

When the user signs in with a phone number, but the account has an associated email, we didn’t prefill the email from the login credentials.

* Actually sign out the user when they cancel giving their password

When a user with email logs in and is asked for a password, they can cancel giving the password. If they do so, we delete the account from the account manager, instead of just selecting another account.

* Do not actually delete the account as this removes history