Fix: iOS: Username color in the conversation messages view show black colour in dark mode  (#5373)
* Fix: Username color to always show a color from the list

By making these changes we now know that when a color is .underfined it will set to .strongBlue.

* ColorScheme: Duplicate value 0.8 to be the first and second value of the array

Remove 0.0 entirely so it doesn't mess up with how we "generate" a UIColor based o our enum.

* Add UnitTests for testing the nameColor method

We test if we get the expectedColor (.strongBlue) when  the accentColor is set to undefined. In both .light and .dark variants

* ColorSchemeTests: Fix test implementation

I wasn't mixing black and white respectively in both test cases.

* Documentation for the tests

* ColorSchemeTests: Improve the naming of the tests

Apply suggestions from code review

Co-authored-by: Bill Chan <billypchan@gmail.com>

* Improve the naming of the tests

Based on the suggestions Katerina made. From: `testForIsStrongBlue_WhenUndefinedInDarkVariant` to `testForIsStrongBlue_WhenAccentColorIsUndefinedInDarkVariant` etc

Co-authored-by: KaterinaWire <57407805+KaterinaWire@users.noreply.github.com>

Co-authored-by: Bill Chan <billypchan@gmail.com>
Co-authored-by: KaterinaWire <57407805+KaterinaWire@users.noreply.github.com>