chore: Integrate SwiftLint into the iOS-Request-Strategy Project #IC-8
Integrating SwiftLint into the iOS Sync Engine project. The .swiftlint.yml file is based on the file that's been added on the wire-ios repository. I've also added a GitHub action for SwiftLint.

shorthand operator has been added to the list because of an error that pops up, so we're going to disable it until we fix it.

line_length is set to 250 because we doesn't seem to hit any errors with that limit. It's the same limit we have in wire-ios project.