fix: Unable to create a conversation with a service SQSERVICES-1394 (#1700)
* fix: check if allowServices is false and set it to true

We need to do that if we want to create a conversation with bot. We need to set the allowServices to true. Then we're going to add service to the accessRoles and the payload would be correct. The reason wasn't working was because we didn't add "service" to the accessRoles so the payload was incorrect.

* fix: remove if statement

* chore: Update version.xcconfig for version: '416.0.1' [skip ci]

Co-authored-by: zenkins <iosx@wire.com>