chore: Remove `teamCreation` flow from the UI Project IC-18 (#5673)
* chore: delete team creation code from the project

the code related to team creation is dead code. It's not being used anywhere in the project that's why we remove it. When and if we want to bring it back we can use git to bring it back.

* chore: delete empty files and groups

* chore: Delete tests related to teamCreation

* chore: Delete reference images for teamCreation

* chore: delete strings that are not in use

* chore: remove more dead code related to teamCreation

* chore: delete and re-organize files

We delete the files that they don't contain any code anymore and we also move some files to better show when the belong in the file structure

* chore: Delete TeamMemberInviteViewControllerSnapshotTests

* chore: delete reference images