fix: Font size for Navigation Bar Items - ACC - 383 (#6039)
* chore: refactor code for rightBarButtonItem

We get rid of two methods and we create a new one where we can use on multiple places. There's no need to have so many methods where the only thing we change is the title of the button

* fix: Update fonts by using createNavigationRightBarButtonItem

We use this method in all places where we need a rightBarButtonItem. We make sure the font size is the correct one

* fix: ChangeEmailViewControllerTests and ChangeHandleViewControllerTests

* fix: showMyDeviceButton font size

* chore: cleanup makeBackButton()

we cleanup the makeBackButton method since we don't need to check for the colors (old vs new) anymore

* chore: make code more readable

Co-authored-by: David-Henner <david.henner@wire.com>

* fix: snapshot tests

Co-authored-by: David-Henner <david.henner@wire.com>