fix: Update colors for NetworkIndicatorView (#63)
Co-authored-by: François Benaiteau <netbe@users.noreply.github.com>

With this PR we update the colors for the NetworkIndicatorView. In addition to that, we get rid of override var preferredStatusBarStyle from these classes since it's not needed anymore.

CollectionsViewController
GiphyNavigationController
ConversationContentViewController
MessageDetailsViewController
FolderCreationController
FolderPickerViewController
GroupDetailsViewController
SettingsClientViewController
UserClientListViewController
ProfileViewController
Last but not least we get rid of two properties in ColorScheme+Helper.swift since they're not used anymore.