chore: CellConfiguration - Get rid of the variant property (#59)
With this PR we get rid of the variant property from the CellConfiguration. We don't need that property anymore since it's the old way of handling colors. This is one of the many PR's that helps us remove deprecated properties and methods from the code base regarding the handling of colors.