chore: add contextMenu and copy to clipboard action for DeveloperMenu cells (#167)
Currently, if we want to copy the datadog id (as an example) and send it to someone, we can't. We either have to send a screenshot or just write down the first couple of characters of the id. That's time-consuming and kind of inefficient.

With this PR we fix this problem by providing a context menu to those cells and the ability for us to copy the contents of the cell.

Co-authored-by: John Nguyen <28632506+johnxnguyen@users.noreply.github.com>