chore: get rid of preferredStatusBarStyle (#210)
Setting preferredStatusBarStyle it's not needed in most places so we get rid of it.

--

Note: There are two places where we still set that property. CallTopOverlayViewController and CallViewController. I'm planning to get rid of those two as soon as we merge the PR regarding the deletion of the old Calling UI