fix: Update failing Tests (#219)
With this PR we update tests that are failing and need to update their reference images. As a result of this PR when running the tests (Wire-iOS) locally from now on you will be able to see 0 test failures. If that's not the case we may have something else that messes things up.