chore: Capitalize only the first letter in a string - ACC-459 (#182)
There's been an issue that's come to our attention where we capitalize every word of a string using the .capitalized. Even though this is the right approach in some cases, in many others it's not.

So we came up with a solution to introduce a new computed property so we can capitalize only the first word of a string. Documentation and tests have been added as well.