fix: Capitalization issues in `SearchHeaderViewController` and `TypingIndicatorView` - SQSERVICES-2034 (#256)
During our accessibility work, we were updating text as well as UI components. During those updates, we started using .capitalized in many strings. That led to capitalizing the first letter of every word of that string. That was the wrong approach.

There is going to be a series of PRs to fix those capitalization issues.

in the case of TypingIndicatorView, the name is uppercased. That's the wrong approach as well. It's not so accessible.

With this PR we're planning to fix the capitalization issues on both TypingIndicator and SearchHeaderViewController.

In #182 we introduced a new property called capitalizingFirstCharacterOnly which capitalizes only the first character of a string.

So we're using that property now to fix the issues mentioned above.