fix: Running the app on Device after merging the CoreCrypto PR (#260)
After merging #255 we discovered an issue where we couldn't run the app on the device. We would get the following error:

Library not loaded: @rpath/WireCoreCrypto.framework/WireCoreCrypto

After investigating with Jacob we realized that we were embedding WireCoreCrypto in Data Model and that doesn't need to be the case. Instead, we embed and sign it in Wire-iOS directly from Carthage, which fixes the issue.