fix: Capitalization issues across multiple views (#262)
After merging #256, this PR will fix some more capitalization issues in multiple views. More specifically, we should avoid capitalizing conversation names across the board. With that said:

We got rid of capitalizing conversation names in:

- `GiphyConfirmationViewController`
- `CanvasViewController`

In addition to that we fixed capitalization issues on different placeholders and hint labels in

- `GiphySearchViewController`
- `CanvasViewController`