fix: There is no pen icon on the preview of an edited message someone is replying to - WPB-3494 (#361)
With this PR we show a the edit icon when someone is replying to a message which has been edited. The problem was that instead of modifying the insets we're adding an empty string + our text attachment. For some reason our icon was still hidden and that solution wasn't working as expected.

So now, instead of an empty string we add the proper left inset and the issue has been fixed.