fix: Transition when presenting a `ProfileViewController` (#372)
In a case where we wanted to tap on someone's profile (to see the profile) and to connect, there was that weird transition that was happening because of TransitionDelegate. TransitionDelegate was the source of the problem. And it is not needed anymore, so it's being removed.

In addition to that ZoomTransition is getting removed as well since we were using it only on TransitionDelegate.

Last but not least we replace viewDidAppear with viewWillApear so we can remove the delay in which the x icon is shown when we present the view