fix: partial search for local users - WPB-3510 (#391)
With this PR I fix an issue where it wouldn't allow for a partial search in terms of names. The problem was how we defined the regular expression within the method. By removing \b.

\b = Word boundary. Matches a word boundary position between a word character and non-word character or position (start / end of string).

So by removing that we allow partial search. I've updated some of the tests to reflect those changes.

The new definition allows for the partial search to work as expected.