chore: Get rid of `LegacyColors` within `SemanticColors` (#392)
With this PR I get rid of the LegacyColors within SemanticColors.

In addition to that ConversationListTopBarViewControllerSnapshotTests I replaced FB's snapshot library with SnapshotTesting library.

Last but not least I cleaned up ConversationSystemMessageCell. Provided a bit of documentation, created a bunch of type alias, and replaced .localized with L10n whenever possible.