chore: Get rid of `ZMSnapshotTestCase` and replace it with `BaseSnapshotTestCase` (#410)
This PR follows #400 where:

Our SnapshotTest classes do not need to inherit from ZMSnapshotTestCase just for setting the contentSizeCategory for fonts. Instead, we create a new Base class: BaseSnapshotTestCase in which we configure the contentSizeCategory.

This is going to be one of many PRs in which I'm planning to replace ZMSnapshotTestCase with BaseSnapshotTestCase

On that note, here are two positive side effects of such a change:

Slowly but surely we're getting rid of snapshot code related to FB's library

Tests are going to be a lot faster since their respective classes are planning to inherit from a much smaller class than ZMSnapshotTestCase.