refactor: InputBarButtonsView for better code readability and interaction with the class (#1024)
With this PR we refactor InputBarButtonsView. It was a class that it was hard to work with. As we try to improve the code base slowly but surely, I took some first steps to improve the ergonomics and the code readability in the class:

Introduced InputBarRowConstants
Introduced documentation for the entire class so its easy to understand how its being used
Separated bigger methods into smaller ones