fix: [iPad] - certificate details screen has the wrong background color (#1201)
With this PR we fix the background color for the Certificate Details view. In addition to that there are two more changes included in this PR.

Replaced old way of setting up fonts, with the new one according to the design guidelines
Introduced a Preview for the view which makes it way easier to observe the view while making changes in real time.