Remove colorSchemeObserver from ZClientVC
When we're for example in Search Conversation screen and we decide to change from light to dark mode or vice versa, the view is being dismissed and we show the conversation view. That can break the flow of the user while in the midst of typing or searching something. In addition to that it doesn't make for a good user experience overall when views are being dismissed for now reason.